add DTO request with validator for register
